Move SCP dependency from scpd.types to scpd.scp
scpd.types is intended to be a client package, while scpd.scp is a server package.
While the server can import the client, the other way around leads to extra dependencies.
